+++ src/onegov.gazette/HISTORY.rst
@@ -1,5 +1,7 @@
 Changelog
 ---------
+1.17.3 (2018-10-30)
+~~~~~~~~~~~~~~~~~~~~
 
 - Uses the latest onegov.form.
   [msom]